feat: add storage select component and move things around
