+++ hyko_sdk/utils.py
@@ -1,4 +1,27 @@
 import re
+from typing import Any, Optional
+
+from pydantic import Field
+
+from hyko_sdk.components import Components
+
+
+def field(
+    description: str,
+    default: Optional[Any] = None,
+    show: bool = True,
+    required: bool = True,
+    component: Optional[Components] = None,
+) -> Any:
+    return Field(
+        default=default,
+        description=description,
+        json_schema_extra={
+            "show": show,
+            "required": required,
+            "component": component.model_dump() if component else None,
+        },
+    )
 
 
 def to_display_name(field_name: str) -> str:
@@ -25,3 +48,25 @@ def to_display_name(field_name: str) -> str:
     readable_name = re.sub(r"(?<!^)(?=[A-Z])", " ", spaced_name).title()
 
     return readable_name
+
+
+mimetype_to_extension = {
+    "text/plain": "txt",
+    "text/csv": "csv",
+    "application/pdf": "pdf",
+    "image/png": "png",
+    "image/jpeg": "jpeg",
+    "image/gif": "gif",
+    "image/bmp": "bmp",
+    "image/webp": "webp",
+    "audio/wav": "wav",
+    "audio/mpeg": "mp3",
+    "video/mp4": "mp4",
+    "video/vnd.avi": "avi",
+    "video/webm": "webm",
+    "video/mpeg": "mpeg",
+    "video/x-matroska": "mkv",
+    "video/quicktime": "mov",
+    "video/x-ms-wmv": "wmv",
+}
+extension_to_mimetype = {value: key for key, value in mimetype_to_extension.items()}
